requirements.txt
python-telegram-bot==20.6
selenium
webdriver-manager
beautifulsoup4
python-dotenv
